# specs/openapi.yaml
openapi: 3.0.3
info:
  title: Приложение для федерации тяжелой атлетики г. Москвы (ФТАМ)
  description: Приложение для спортcменов, тренеров, организаторов и представителей ФТАМ - автоматизация деятельности ФТАМ.
  version: 1.0.11

# servers:
#   - url: https://wlmoscow
...
paths:
...
  /tcamps:
    get:
      tags:
        - tcamps
      summary: Получить все сборы
      operationId: getTCamps
      parameters:
        - name: date
          in: query
          description: Дата начала сборов
          required: false
          schema:
            type: string
            format: datetime(utc)
        - name: city
          in: query
          description: Город
          required: false
          schema:
            example: "Москва"
            type: string
        - name: sort
          in: query
          description: Сортировка. Список имен атрибутов, разделенных запятыми, для сортировки. Если перед именем атрибута стоит знак минус, то сортировка производится по убыванию, в противном случае — по возрастанию.
          required: false
          schema:
            example: date,-city
            type: string
        - name: page
          in: query
          description: Пагинация - номер страницы.
          required: false
          schema:
            example: 3
            type: integer
        - name: batch
          in: query
          description: Пагинация - количество элементов на странице
          required: false
          schema:
            example: 10
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllTCamps'
        '500':
          description: Internal Server Error
    post:
      tags:
        - tcamps
      summary: Организация сборов
      operationId: createTCamp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTCamp'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TCamp' 
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
  /competitions:
    get:
      tags:
        - competitions
      summary: Получить все соревнования
      operationId: getComps
      parameters:
        - name: name
          in: query
          description: Название мероприятия
          required: false
          schema:
            type: string
            example: "Кубок Москвы"
        - name: date
          in: query
          description: Дата начала (возможно только год)
          required: false
          schema:
            type: string
            format: datetime(utc)
        - name: city
          in: query
          description: Город
          required: false
          schema:
            example: "Москва"
            type: string
        - name: age
          in: query
          description: Возрастная группа
          required: false
          schema:
            example: "Юниоры, юниорки"
            type: string
        - name: sort
          in: query
          description: Сортировка. Список имен атрибутов, разделенных запятыми, для сортировки. Если перед именем атрибута стоит знак минус, то сортировка производится по убыванию, в противном случае — по возрастанию.
          required: false
          schema:
            example: date,-city
            type: string
        - name: page
          in: query
          description: Пагинация - номер страницы.
          required: false
          schema:
            example: 3
            type: integer
        - name: batch
          in: query
          description: Пагинация - количество элементов на странице
          required: false
          schema:
            example: 10
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllComps'
        '500':
          description: Internal Server Error
    post:
      tags:
        - competitions
      summary: Организация соревнований
      operationId: createComp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComp'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comp'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
  /users/signup:
    post:
      tags:
        - users
      summary: Создание пользователя
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: Internal Server Error
  /users/login:
    post:
      tags:
        - users
      summary: Вход в аккаунт
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUser'
        required: true
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                  expires_in:
                    type: integer
                    example: 3600
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: 123e4567-e89b-12d3-a456-426655440000
                      email:
                        type: string
                        format: email
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
  /users/{id}:
    get:
      tags:
        - users
      summary: Получить пользователя по ID
      description: ''
      operationId: getUserByID
      parameters:
        - name: id
          description: ID пользователя
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
        '500':
          description: Internal Server Error
    put:
      tags:
        - users
      summary: Обновить email и пароль пользователя
      description: Для авторизованных пользователей.
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
    delete:
      tags:
        - users
      summary: Удалить пользователя
      description: Для авторизованных пользователей.
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successful operation
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal Server Error
  /users/{id}/competitions:
    get:
      tags:
        - users
      summary: Получить все соревнования организатора
      operationId: getCompsByOrgId
      parameters:
        - name: id
          in: path
          description: ID организатора
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllComps'        
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /users/{id}/tcamps:
    get:
      tags:
        - tcamps
      summary: Получить все сборы организатора
      operationId: getTCampsByOrgId
      parameters:
        - name: id
          in: path
          description: ID организатора
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllTCamps'          
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
components:
  schemas:
    TCamp:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        city:
          type: string
          example: Москва
        address:
          type: string
          example: Московская 1
        beg_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
    Comp:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        name:
          type: string
          example: Кубок Москвы
        city:
          type: string
          example: Москва
        address:
          type: string
          example: Московская 1
        beg_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        age:
          type: string
          example: Юниоры, юниорки
        min_sports_cat:
          type: string
          example: КМС
        antidoping:
          type: boolean
          example: true
...
    CreateComp:
      type: object
      properties:
        name:
          type: string
          example: Кубок Москвы
        city:
          type: string
          example: Москва
        address:
          type: string
          example: Московская 1
        beg_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        age:
          type: string
          example: Юниоры, юниорки
        min_sports_category:
          type: string
          example: КМС
        antidoping:
          type: boolean
          example: true
    CreateTCamp:
      type: object
      properties:
        city:
          type: string
          example: Москва
        address:
          type: string
          example: Московская 1
        beg_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
...
    AllTCamps:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/TCamp'
    AllComps:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/Comp'
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        email:
          type: string
          format: email
        role:
          type: string
          example: Coach
        user_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440001
    UpdateUser:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          example: '12345'
    CreateUser:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          example: '12345'
        role:
          type: string
          example: Coach
    LoginUser:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          example: '12345'
        

  # securitySchemes:
  #   petstore_auth:
  #     type: oauth2
  #     flows:
  #       implicit:
  #         authorizationUrl: https://petstore3.swagger.io/oauth/authorize
  #         scopes:
  #           write:pets: modify pets in your account
  #           read:pets: read your pets
  #   api_key:
  #     type: apiKey
  #     name: api_key
  #     in: header